+++ firmware/CHANGELOG.md
@@ -27,11 +27,21 @@ All notable user-facing or behavior-altering changes will be documented in this 
 
 ## XXX 2021 Release
 
+### Added
+ - USB Mass Storage: The connected SD card will be mounted over USB if rusEFI connected to a PC via USB.
+ - GM 60/2/2/2 trigger https://github.com/rusefi/rusefi/wiki/All-Supported-Triggers#gm
+ - TriTach trigger https://github.com/rusefi/rusefi/wiki/All-Supported-Triggers#tritach
+ - Skoda Favorit trigger https://github.com/rusefi/rusefi/wiki/All-Supported-Triggers#skoda-favorit
+ - Add fallback logic handling failed MAP sensor.  In case of failed MAP, ses either a fixed MAP value, or a table that estimates MAP based on TPS and RPM.
+
+### 2021 Printing Ink Day
+
 ### Breaking Changes
  - Closed loop idle timing behavior changed to no longer fall off control as entering/leaving the idle region. It now sharply engages/disengages upon entering/leaving the idle area.
  - Idle phase logic uses the same idle detection thresholds as the main idle controller instead of its own thresholds.
 
 ### Added
+ - Main Relay bench test button
 
 ### Fixed